backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
google-generativeai==0.3.1
PyPDF2==3.0.1
python-docx==1.1.0
aiofiles==24.1.0
python-dotenv==1.0.1
reportlab==4.2.5
bcrypt==4.2.0
